+++ NEWS.md
@@ -1,5 +1,8 @@
 # centr 0.2.4
 
+* `mean_center()` and `median_center()` now work even if the geometry column is not
+  named geometry (#20).
+
 # centr 0.2.3
 
 * Adjust `testthat` tolerance to 4 decimal points due cran checks.